+++ app/serializers_fast/fast_activity_serializer.rb
@@ -1,0 +1,6 @@
+class FastActivitySerializer
+  include FastJsonapi::ObjectSerializer
+
+  set_type :activity
+  attributes :title, :description
+end