+++ config/routes.rb
@@ -80,6 +80,7 @@ Rails.application.routes.draw do
       resources :events, only: %i[index] do
         get :filter_values, on: :collection
       end
+      resources :activities, only: %i[create]
     end
   end
 
@@ -120,6 +121,7 @@ Rails.application.routes.draw do
       get :decline, on: :member
     end
     resources :deliveries, only: %i[new destroy]
+    resources :activities, only: %i[new]
     resources :guild_invites, only: %i[] do
       get :find, on: :collection
     end