+++ frontend/packs/activity_form.jsx
@@ -1,0 +1,4 @@
+import ActivityForm from "components_react/activity_form/activity_form";
+import WebpackerReact from "webpacker-react";
+
+WebpackerReact.setup({ActivityForm});